+++ src/lib/remark-plugins/remark-plugin-adjust-link-urls/__tests__/remark-plugin-adjust-link-urls.test.ts
@@ -2,7 +2,7 @@ import {
 	expandUrlTestCasesWithParams,
 	TestCase,
 } from 'lib/testing/expand-url-test-cases-with-params'
-import { preAdjustUrl } from 'lib/remark-plugin-adjust-link-urls'
+import { preAdjustUrl } from 'lib/remark-plugins/remark-plugin-adjust-link-urls'
 
 const testEachCase = (testCases: TestCase[]) => {
 	const allCases = expandUrlTestCasesWithParams(testCases)
